feat: 오늘의 한마디 페이지 컴포넌트 분리 및 모듈화 구조로 구현
