+++ src/pages/today-words/constants.ts
@@ -1,0 +1,35 @@
+import type { Message } from './types';
+
+export const dummyMessages: Message[] = [
+  {
+    id: '1',
+    user: 'user.01',
+    content:
+      '공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다.',
+    timestamp: '2024.04.30',
+    timeAgo: '11시간 전',
+  },
+  {
+    id: '2',
+    user: 'user.01',
+    content:
+      '공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다.',
+    timestamp: '2024.04.30',
+    timeAgo: '11시간 전',
+  },
+  {
+    id: '3',
+    user: 'user.01',
+    content: '공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다.',
+    timestamp: '2024.04.30',
+    timeAgo: '12시간 전',
+  },
+  {
+    id: '4',
+    user: 'user.01',
+    content:
+      '공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다. 공백 포함 글자 입력입니다.',
+    timestamp: '2024.04.30',
+    timeAgo: '11시간 전',
+  },
+];
