+++ src/pages/today-words/types.ts
@@ -1,0 +1,7 @@
+export interface Message {
+  id: string;
+  user: string;
+  content: string;
+  timestamp: string;
+  timeAgo: string;
+}
